requirements.txt
pytest==7.4.3
playwright==1.40.0
pytest-playwright==0.4.3
allure-pytest==2.13.2